Move CallCommand into it's own package/file
* Move internal/version.go into internal/version/version.go
* Move CallCommand from pkg/model/model.go into internal/command/command.go
  as func Call()
